ENH: count_neighbours picks algorithm depending on cumulative
If cumulative is True (default) the old algorithm is used,
where any rbin further than the pairs are updated up on a pair is
visited, producing a cumulative result.

If cumulative is False, the new algorithm is used,
where a rbin is updated if and only if the pairs fall into that bin,
producing a differentiation result.