CS matrix prune method should copy data from large unpruned arrays (#6623)
* BUG: CS matrix prune should copy data from larger arrays

Pruning resizes the "data" and "indices" arrays after explict zeros
have been removed from a sparse matrix. However, this is implemented
using slicing, which just returns a view of the original array. If
the original array is relatively large, a new array should be used
instead of a view, allowing the original to be garbage collected.

* Add self to list of contributors