io/matlab/mio.py: handle objects supporting PEP 519
We don't need to specifically check for a `__fspath__` attribute here, since
part of PEP 519 is enhancing `__builtins__.open` to support path-like objects.

The logic changes are quite small: first, try to read from whatever is given to
this function. If that succeeds, it's already a file-like object, and it's
returned as-is. If it's not readable, try passing it to `open`. If that
succeeds, return the new file-like object. If it fails with an IOError, check
for object type, append '.mat' if appropriate, and try calling `open` again.

If the user passes something nonsensical to `_open_file`, `open` will throw a
`TypeError`, which is a perfectly sane thing to do:

"""
Python 3.6.0rc1 (default, Dec  8 2016, 17:06:38)
[GCC 5.4.0 20160609] on linux
Type "help", "copyright", "credits" or "license" for more information.
>>> open(object())
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
TypeError: expected str, bytes or os.PathLike object, not object
"""