MAINT: fix signal.resample for small input arrays
The current version of resample(x,n) fails when the smaller of x.size
and n is even. With real-valued x, the problem appears only on
downsampling (for mathematical reasons), with complex-valued x, both on
upsampling and downsampling.

Simple test cases:

1. For x=[1.+0.j, 0.+0.j], resample(x,4) should produce [1.+0.j,
0.5+0.j, 0.+0.j, 0.5+0.j], but does not.

2. For x=[1., 0.5, 0., 0.5], resample(x,2) should produce [1., 0], but
does not. The problem lies with the Fourier component at frequency N/2.
Say we want to upsample from an even N to a larger num. The current code
takes the single Fourier component at N/2 in the input spectrum and
assigns it entirely to num-N/2 in the output spectrum. Instead that
should be split evenly between frequencies N/2 and num-N/2. Vice versa
say we want to downsample from Nx to a smaller N that is even. The
current code takes the Fourier component at Nx-N/2 in the input spectrum
and assigns it to N/2 in the output spectrum. Instead we need to add to
that the Fourier component in the input spectrum at N/2.

The revision treats these special cases properly.

Added tests of resample() by the fft method

In scipy v.0.18.1 resample() fails when the smaller of the input and
output array size is even. These tests expose that malfunction.

Edited doc strings for fft functions

1) Some version of this appears repeatedly in help(fft), help(rfft),
help(fftn): Note that ``y(-j) == y(n-j).conjugate()``

This is wrong in two ways:
    a) y(-j) is actually defined as y(n-j).
    b) perhaps the author meant to write ``y(j) == y(n-j).conjugate()``.
    But this is true only if the input x is a real-valued function, yet
    the statement appears even under complex fft. Better to omit this.

2) Under help(fft) we get this:
    "For `n` even, ``A[n/2]`` contains the sum of the positive and
    negative-frequency terms.  For `n` even and `x` real, ``A[n/2]``
    will always be real."
This is misleading: There is only one frequency term at n/2, not a
positive and a negative frequency. The statement gives the impression
that n/2 is somehow handled differently from the other frequencies.
Better to omit this.

Update basic.py

Restored and clarified some lines in doc strings for fft() and fftn()
regarding the special case where input array is real-valued.