BUG: Fix signal.dfreqresp for SS systems
Unfortunately, `dfreqresp` used the easily confused names `system` and
`sys` to refer to the unmodified input system and potentially converted
to ZPK system respectively. A typo then led to the ability to fall out
of an if block without actually computing the output (see gh-7335). This
fixes the logic for SS systems, avoids unnecessary logic for converting
from SS to ZPK (since ss2zpk just converts to TF form internally), and
adds a test for `dfreqresp` on an SS system.