BUG: 1d ndimage.affine_transform endianness fix
Resolves #7388. PR #4150, which fixed, #4127 handled
ndimage.geometric_transform, ndimage.map_coordinates, and the branch of
ndimage.affine_transform handling multidimensional input, but missed the 1d
case.

Underlying problem: when some ndimage functions are passed an ndarray to write
output into, the writing is done by C code that is unaware of dtypes and just
uses the system's native endianness. We work around this by applying a
decorator that byteswaps the output if it has a non-native dtype.