Added a sanity check to signal.savgol_filter (#7596)
* Added a sanity check to the savgol_filter

ENH: added a sanity check to the scipy.signal.savgol_filter

When using savgol_filter with mode's default parameter "interp", passing
an x array with a size less than the `window_length` results in a rather
cryptic TypeError raised back in numpy.polyfit with the message
"expected x and y to have the same length". This message offers no
indication that the x size in relation to the window_length is causing
the issue. I've included a value check with a more informative error
message. The check piggy backs off of the `mode == "interp"` code flow,
as it only affects that mode.

* Removed some redundant parentheses

* Added backticks around mode and x, as requested.

* Changed double quotes to single

Docstrings use single quotation marks for signalling strings.