TST: convert yield tests to plain tests
Change yield tests to plain tests, where there is no strong reason to
keep the test parametrized.

Typically, with nose, the yield tests were used to indicate for which
tested values a test failed. However, pytest can print local variables
and is configured to do so in scipy.test, so this use case is no longer
necessary so the tests can be written in a simpler way.