BUG: io: fix confusion in FortranFile about fortran order in multidim arrays
The test files used in reality contain data in C-order, so rename them
to match. For example, the old "fortran-si4-5x3x1.dat" could be read
correctly only by::

    integer :: a(3,5)
    open(1, file='fortran-si4-5x3x1.dat', form='unformatted')
    read(1) a

That the dimensions were transposed was also visible in
test_fortranfiles_read, which did `reshape(dims)` instead of
`reshape(dims, order="F")` to unravel the 1D data read.

To retain backward compatibility, do not try to deal with Fortran-order
issues, and leave them completely to the user.