ENH: added sygst/hegst routines to lapack
This enables the sygst and hegst routines for converting a
generalized eigenvalue problem into a standard one.

This will effectively allow one to use _all_ the eigenvalue problem
routines by chosing the appropriate method.

Tests are added to check against the generalized problems.

It is unclear to me why a rtol <1e-5 is necessary. The LAPACK sources
shows the exact same steps as I inserted. However, the numerical
accuracy is lacking?

Signed-off-by: Nick Papior <nickpapior@gmail.com>