MAINT: special: Mark a test that fails on i686 as a known failure.
The calculation `special.fdtri(1, 1, 0.5)` sometimes returns nan on i686
systems.  Before the change in https://github.com/scipy/scipy/pull/7919,
the test suite had this comment:

    # cephes.fdtri(1,1,0.5)  #BUG: gives NaN, should be 1

but there was no issue for tracking the problem.  That test was restored
in gh-7919, and it passes on 64 bit architectures.

This change makes that calculation a separate test that is marked as a
known failure.  The github issue for this test is now gh-7931.