TST: test_falker not robust against numerical fuss in eigenvalues
The generalized eigenvalue-problem with two matrices A and B
as given in the test_falker is such that one generalized eigenvalue
if infinite, one is a nerative real, two are exact zeros, and
the remaining two are complex conjugates of each other with a negative
real part.

To compare whether eigenvalues obtained from `eig` and from `eigvals` are
the same, the code compares sorted finite eigenvalues.

The sorted vector of eigenvalues is sensitive to numerical fuss in real parts
of the complex conjugate eigenvalues.

It may happen, with a change from one solving library to another, or even with
the version upgrade of that library, that the ordering of the result of `eig`
be `(-real + 1j*imag, -real -1j*imag)`, while the oridering of the result of
`eigvals` be `(-real - 1j*imag, -real + 1j*imag)`, which causes the test to fail,
even though the result is correct within the numericl precision.

To fix this issue, this PR introduces routine `clear_fuss` which zeros out
certain number of trailing binary digits (default is arbitrarily chose as 7) in a real
number's mantissa. For complex numbers fuss is independently cleared for real and
imaginary parts.

Ordering permutation is computed for the vector where the fuss has been cleared,
and the permutation is applied to the actual vector of eigenvalues.  These permuted
vectors are then compared with `assert_all_close`.

STY: pep8 fixup