ENH: stats: Use formulas for the MLE fit of the exponential distribution.
Exact formulas for the maximum likelihood estimation of the parameters of
the exponential distribution are known, so use those instead of using
numerical optimization to solve for them.

There is a change in behavior for one edge case.  In a call such as
    expon.fit(x, floc=0)
if any value in x is less than 0, a FitDataError exception will be raised.
Before this change, the scale returned by that call would be, in effect,
the sum of the nonnegative values divided by len(x).  For example,
    expon.fit([-1, 1, 3, 5], floc=0)
would return (0, 2.25).  If the values outside the support were to be
completely ignored, the value for the scale in that case should be the
mean of the nonnegative values, 3.0.  So now, instead of returning
a questionable result, the code simply doesn't allow values outside
the support.