ENH: stats: Use formulas for the MLE fit of the lognormal distribution.
When the location parameter is fixed, exact formulas for the maximum
likelihood estimation of the parameters of the lognormal distribution
are known, so use those instead of using numerical optimization to solve
for them.

There is a change in behavior for one edge case.  In a call such as
    lognorm.fit(x, floc=0)
if any value in x is less than 0, a FitDataError exception will be raised.

The expected value for an estimated shape in the test
'TestFitMethod.test_fix_fit_2args_lognorm' in test_distribution.py
is now computed with the explicit formula.  The old value was inaccurate.