ENH: optimize: Always update Hessian / reduce min_curvature default
Two improvements:

1. Reduce `min_curvature` default for BFGS `skip_update`

The default value of `min_curvature` was causing errors in some
situations. Too many steps were getting rejected and this was
causing the solver to fail. I notice this when trying to modify the
HessianUpdateStrategy.

2. Always update Hessian when using HessianUpdateStategy

This approach is more solid because it keeps improving hessian
estimation even for failed iterations.