ENH: Add _argmaxima1d for maxima detection in 1D arrays (#8350)
* ENH: Add _argmaxima1d for maxima detection in 1D arrays

This Cython function is intended to be used for a peak finding function
proposded in #8264. Exposure to the user is not intended. It is meant as
an alternative to scipy.signal.argrelmax` and compares as follows:

- It only works on 1D arrays
- It should be faster for all use cases
- It can detect flat maxima (more than one sample wide).

Unit tests are provided as well.

* ENH: Use memoryview for input of _argmaxima1d

- Memoryview seem to be the preferred way to work with input arrays in
  Cython. This change gives (in some cases) a performance boost but is
  not significant. The created array `maxima` still uses the old syntax
  because its content needs to be returned to Python as an array
  anyways.
- Fixed tests to deal with new behavior because of the above change.
- Added short comparison to `scipy.signal.argrelmax` in Notes section.

Addresses reviews in PR #8350.

* TST: Use Python 2.7 compatible match string for exception checking

The old match-string doesn't seem to work in Python 2.7 and 3.4. This
hopefully addresses all standing CLI issues (see PR #8350).