ENH: sparse.linalg: implement atol option for iterative fortran solvers
Add a way to specify absolute tolerance in iterative fortran solvers.

Previously, these solvers had only the 'legacy' behavior: `tol` is the
tolerance relative to norm(b) and also considered as the absolute
tolerance on the first iteration (!). The behavior originates
from the checks of the form

    IF ( <rc>NRM2( N, WORK(1,R), 1 ).LT.TOL) GO TO 200

This commit changes the internal workings of the fortran solvers
so that:

- The solvers internally deal only with the absolute 2-norm convergence
  criterion. The threshold for this criterion is computed before entering
  the solver.

- All BNRM2 computations are moved python-side.

- STOPTEST2 is removed, as it is easily written in python.

- The strange 'legacy' behavior of the tol= argument is preserved, but
  with a deprecation warning.

- Change the first-iteration check in the fortran codes to check for
  ``RESID .LE. TOL`` instead of ``RESID .LT. TOL`` so that zero rhs
  vectors are handled correctly also for tol=0.