BUG: sparse.linalg/gmres: fix termination condition when using preconditioner
Make changes to ensure tol= and atol= kwargs in gmres() apply to the
residual rather than the preconditioned residual.

Be explicit in the GMRESREVCOM that tolerances apply to the
left-preconditioned residual.

When inner loop residual decreases below tolerance in the GMRESREVCOM,
break out from the inner loop (similarly as during restart) rather than
exiting with success. This ensures _stopcheck is run, which checks the
termination condition for the non-preconditioned residual before
exiting.

As the termination conditions for the outer and inner loops may
mismatch, control the inner tolerance adaptively.