BUG: optimize: fix Bland's Rule for pivot row/leaving variable, closes #8561 (#8568)
Choosing the first of the possible pivot rows (as it does now) is not
in accordance with Bland's Rule, since the basis is not sorted by
variable index. This results in some inputs cycling/not terminating when
Bland's anti-cycling rules are used (see issue). This fix does a lookup in
the basis array to make sure the correct row is chosen.