ENH: Cythonize peak_widths (#8594)
* ENH: Cythonize inner loop of peak_widths

The actual algorithm for peak width calculation was moved to Cython
which allows a more straight forward approach using loops. This has the
following advantages:
- Significant performance gains especially for large input arrays.
- The GIL can be released during calculation.

Other notable changes:

Add some validation for `prominence_data` which is advantageous if
supplied through the user and necessary for bounds checking.

Use the test signal from peak_prominences for in the docstring example
which might be clearer for users.
Reworded description of arguments, raised exceptions and Notes.

* TST: Update tests for peak_widths

- Add test for non-contiguous case.
- Test exceptions for mismatching peak and prominence data.
- Make assertions more tolerant because of numeric errors when
  calculating the interpolated width.
- Use pytest.mark.parametrize where applicable
- test exception when argument prominence_data contains None.
- Adjust exceptions matches after rebase. The function peak_prominences
  is called within peak_widths which raises an excpetion for invalid
  peaks.

* MAINT: Improve input validation and corresponding tests

The shape of prominence_data is now properly checked against the shape
of supplied peaks. Tests to protect against regressions where added as
well.

Furthermore "not None" is used in favor of "@cython.nonecheck" as they
have a slightly different behavior. The latter only checks for None
during attribute access while the former ensures that None isn't valid
input. "not None" is prefered because it fails early and with a clearer
error message.

Removed usage of @pytest.mark.parametrize in favor of internal loops
because "data sets should not appear as arguments to parametrize".

* TST: Test rules for finding an intersection

This basically covers the edge cases when the signal contains one or
more samples with the exact evaluation height in which case they should
be counted as valid intersections. As a consequence flat peaks always
have a width of 0 for rel_height=0.