BUG: stats: Fix logpdf method of invwishart.
The private function _cho_inv_batch in _multivariate.py computes
the inverse of a sequence of symmetric positive definite matrices
using the Cholesky factorization. There was a bug in this function,
where it didn't correctly copy the lower part of the result to the
upper part to make the resulting inverse(s) symmetric.

Closes gh-8844.