TST: Test cases for `as_euler` with extrinsic rotations
These test cases compare the actual angles returned by the algorithm. It appears that we cannot always guarantee that these will be th same as the input angles, only that the final rotation generated by the angles will be the same.

For intrinsic rotations, it appears that the algorithm never gives the same angles.