BUG: Correct "remez" error message (#9066)
* Correct "remez" error message

The error case for remez is erroneously reporting the max iterations has been achieved, this is not the case.
The actual number of iterations is less and the failure is due to a decreasing deviation on an early iteration. This is due to the error being too small - smaller than round-off - and indicates specifications that are too lax. A more accurate error message indicates that the number of iterations is less than maximum, and that it might be averted by increasing the transition band width.

* Update sigtoolsmodule.c

Move "int niter" up to top of function.