MAINT: Add warnings if pivot value is close to tolerance in linprog(method='simplex') (#9081)
* Add contribution to THANKS.txt

* Refactor pivot technique into new function.

* Add warning if pivot value is close to the tolerance.

The pivoting procedure may produce pivot values very close to
zero. Dividing by these values can cause the simplex method
"blow up". Adding the warning notifes the user if this is likely
to occur and provides some hints to prevent this.

* Update tests to reflect warnings.

* Fix comment on expected behaviour of test.

* Slight changes to codestyle and warning message.

Removes additional whitespace from tests in test_linprog.
Changes the order of suggestions in the previously added warning
message. The original message firstly suggested trying to increase
the tolerance. This may work in theory, but shouldn't be the first
choice.