ENH: add capability to specify orders in pade func (#9211)
* ENH: add capability to specify orders in pade func

Currently interpolate.pade only takes the order of the denominator <m>
with the order of the numerator assumed from the order of the inputted
array <an>. Added an optional parameter <l> (following standard naming
convention of the pade function) which specifies the order of the
numerator. This is useful if the order of the inputted array returns
a higher order (and computationally more expensive) numerator than needed.
This change does not affect any currently written code.