BUG: Fixed nested use of SLSQP by storing correct internal state.
See issue #9316.

The Fortran SLSQP implementation made use of local variables with the
SAVE attribute. The state of these variables were not correctly restored
if nested calls to SLSQP were made. The problem was fixed by passing all
these variables in as parameters to SLSQP, instead of relying on the
SAVE attribute. In this way, the correct state for different SLSQP
optimizations that are running simulataneously or nested can be stored.