ENH: integrate: quad() for vector-valued functions
Implement a variant of the QUADPACK integration algorithm in Python, and
use it for integrating vector-valued functions.

This is useful in cases in which the function is not easy to calculate
elementwise, or a different norm than the max-norm is wanted for error
estimation, in which cases running quad() separately for each element is
not optimal.

The vectorized algorithm also allows for some degree of parallelization.