BLD: Modify rv_generic._construct_doc to print out failing distribution name.
When a distribution's docstring fails to build, either when building the
refguide or during import of scipy.stats, a TypeError is thrown without
any indication of the offending distribution.
The situation can be triggered by placing %-signs in an distribution's class
docstring, perhaps as part of an example, and these do not get handled by
doccer.docformat().
This change replaces the original Exception with a new one whose message is
"Unable to construct docstring for distribution <NAME>:  TypeError('not ...')"
which gives enough information to diagnose the problem.