DOC: Copy solver description to solver modules
The linear programming solver is currently split between 5 files
a top level interface, a utility library and a module for each of
the three solvers. With the addition of the revised simplex method
the documentation of the individual methods has been improved in the top
level interface. This PR copies the improved documentation into the
module docstrings of each solver module (with light modifications).