ENH: Enhance efficiency of some stats.kstwo basic operations.
Change stats.kstwo's momtype from the default of 1, which uses self.ppf(),
to 0, which uses self.pdf() when computing the distribution's moments.
Added kolmogni(n, q) in _ksstats.py to calculate the ppf/isf for kstwo.

Fixed a general bug in rv_continuous._ppf_single(): If self.a is 0, it
incorrectly used -10 as the left-hand endpoint of the seach interval passed to
optimize.brentq.
If a distribution is excluded from the basic entropy calculation in
test_continuous_basic.py, also exclude it from the private_entropy, vec_entropy
and entropy_vect_scale entropy calculations.

Lower n=100 to n=10 in the example parameters for kstwo in _distr_params.py.
Inside TestKSTwo:
- Add test_isf_of_sf() and test_isf_of_sf_sqrtn() methods.
- Keep CDFs and SFs away from 1 when doing round-trip cdf/ppf sf/isf tests, as
these are subject to more error due to the extreme flatness of the function at
the end-points. (Computing the sf(isf(n, 1-p, n), n) for a CDF probability p
close to 1 works fine, as does cdf(ppf(1-p, n), n).)

All this reduces the computation time in in the kstwo tests by a factor of ~10.