MAINT: special: Fix a warning generated when building the AMOS library.
Unlike the other ZBES* functions, in ZBESY the variable R1M5 is not
declared, so its type defaults to REAL*4 (i.e. single precision float).
This means we get a warning when zbesy.f is compiled, because D1MACH
returns a double precision value:

    scipy/special/amos/zbesy.f:183:13:

           R1M5 = D1MACH(5)
                 1
    Warning: Possible change of value in conversion from REAL(8) to REAL(4) at (1) [-Wconversion]

Presumably this is a mistake.  The fix is to add the declaration of
the variable R1M5 as double precision.

(I can hear the Fortran coders out there tsk-tsking and saying
"`IMPLICT NONE` would have caught that!".)