MAINT: Simplify shape checks in linprog clean_input
The `b` arrays are expected to be 1D arrays with the same number
of elements as there are variables. This commit combines the
checks for the number of dimensions and number of variables by
checking the shape directly.Further this commit removes
explicitly checking ``len(c)`` in favour of the variable ``n_x``.