TST: special: only show max atol/rtol for test points that failed
Currently when using `FuncData` the maximum atol/rtol for all points
is printed when there are only some bad points. This makes it hard to
diagnose what went wrong because e.g. the atol for the entire test set
could be quite high but acceptable (because there are some very large
values being tested). Fix that by only showing the maximum atol/rtol
within the points that failed.