use randn and reduce # of vectors in test_tolerance_float32
In the failure, lobpcg randomly converged to (got stuck at) the eigenvalue 5 instead of 4. This is not a bug - this is a feature, of lobpcg. So I change the test to be a bit more favorable to lobpcg, starting with randn vectors, rather than rand, and reduce # of vectors from 4 to 3, to give lobpcg more space for iterations.