MAINT: ndimage: More robust check for output being a numpy dtype (#10883)
The current logic assumes that all classes are types (which is only
semi true, but probably OK). It also assumes that all numpy dtypes
are not only instances of `np.dtype` but that their type:
`type(dtype) is np.dtype`.
This replaces both checks with the corresponding, more robust,
`isinstance` check.

Note that this is currently not relevant, but will become so if
NumPy redesigns the dtype system, which would (in all likelyhood)
encompass the creation of `np.dtype` subclasses.