Avoid undefined behaviour on float to unsigned conversion
Routines in ndimage rely on undefined behaviour when converting from
the result stored in a double float variable to an unsigned integer
value, i.e. one of npy_bool, ubyte, ushort, etc.  x86 happens to give
the correct negative value in integer form but the aarch64 equivalent
fcvtzu is not so forgiving and sets the result to zero on all
undefined input.  As a result test_ndimage.py fails on aarch64.

This fix adds a new macro CASE_FILTER_OUT_SAFE that inverts the sign
of the double before converting negative numbers to unsigned and
finally negates the integral value again to give the desired result.
Both those operations are well defined.