MAINT: revamp setup for maximum bipartite matching benchmarks
The setup code used in the previous version of the benchmarks was
awfully slow compared to the matching code itself. As such, we opt
to just use NumPy to create sparse matrices instead. The downside
is that we no longer have the exact density we want, but as far as
benchmarking is concerned, that's not really an issue.

In the process, we change the actual sizes and densities to ensure
that we are solving non-trivial matching problems, and to obtain
problems that run in the millisecond range. More precisely, the
graphs we create have the number of edges and maximum matchings
given as follows:

n     | density | edges | size of matching
----- | ------- | ----- | ----------------
 5000 | 0.0001  |  2499 |              171
 5000 | 0.0005  | 12493 |             4346
 5000 | 0.001   | 24982 |             4963
 7500 | 0.0001  |  5625 |             3416
 7500 | 0.0005  | 28113 |             7261
 7500 | 0.001   | 56214 |             7496
10000 | 0.0001  |  9999 |             5418
10000 | 0.0005  | 49989 |             9922
10000 | 0.0001  | 99945 |             9999