MAINT: mypy: remove mpmath from the ratchet
Mpmath doesn't have types, is unlikely to get types, and, since it's a
test-only, optional dependency we shouldn't rely on users having its
types installed even if they did exist. For that reason, remove it
from the ratchet and add `type: ignore[import]`s on Mpmath imports so
that the module is typed as `Any`.