ENH: Test invalid job? combinations for ?gejsv
The previous PR extends the wrappers tests to include complex matrices.
When jobv(or jobu) = 'W' and jobt = 'T' and m = n then U is computed
in V as the  right singular vectors of A^* and copied back to the U
array.

The test now detects invalid combinations are provided and tests that
the correct exit code is provied by info.