ENH: Add wrapper for ?pteqr (#11665)
* TST: Add initial tests for lapack ?pteqr

This commit addds initial test code for pteqr as well as the real case from the NAG manual example.

* TST: Add complex NAG example for pteqr

This commit adds the complex exmaple from the nag manual to the pteqr test. It also reorders the parametrization.

* TST: Add tests for non-zero info, adhere to PEP8

* TST: pteqr add tests for compz = V
This commit adds tests for when compz=V by creating random matrix A and then finding the tridiagonal form with ?sytrd. It modifies the NAG test to use ?hetrd with the supplied A.

* TST: change compz=V A construction

this commit changes the way A is constructed for compz=V. It builds Hermitian A from Q**T * tri * Q = A by creating Q and tri.

* TST: remove NAG tests that require missing routines

* ENH: linalg: Add Python bindings for ?pteqr

* TST: Remove branching, add z inputs

Adds use of z for all compz cases. All random generated tests pass

* TST: break up pteqr megatest

breaks up pteqr megatest, new function to create pteqr parameters for each test, NAG manual ignore sign of z

* TST: loosen tolerance to e-13

* MAINT: linalg: in dpteqr, ldz rules depend on compz

* MAINT: linalg: fix test_lapack error introduced in rebase

* MAINT: linalg: tweaks per comments on gh-11657

* MAINT: linalg: rename `compute_z` per convention, hide undocumented `work`

* TST: lapack: rename `compute_z` in tests, `work` no longer returned

* MAINT: linalg: change ?pteqr compute_z argument from char to int

* MAINT: linalg: fixed minor bugs in pteqr wrapper

* MAINT: linalg: minor fixes per reviewer comments

* THANKS.txt addition

Co-authored-by: Sam Wallan <44255917+swallan@users.noreply.github.com>
Co-authored-by: kai-striega <kaistriega@gmail.com>