DOC: signal: Unify lsim and lsim2 examples.
This change makes the Examples sections of lsim and lsim2
identical.  I tweaked the numbers a bit, and I lowered the
order of the filter to 5.  lsim generates a warning when
the order is 6.

I also edited the double integrator example to make clear
that we're using the state space representation of the
linear system.