DOC: interpolate: degree is not order in splines (#12109)
The following explanation of parameters said `k` is the B-spline order, while here it uses `degree`, but `degree` is different from `order`, and they satisfy `order = degree + 1`. The same issue https://github.com/scipy/scipy/pull/11861 has been corrected.

* correct previous commit

Sorry, I think I had made a mistake. Here `k` is degree, not order. So the description of `k` in the Parameter should be corrected, not the sentence above the Parameters.