ENH: More helpful/forgiving io.wavfile errors (#12110)
* ENH: More helpful io.wavfile.read error messages

Print what bytes were actually seen, so the user has more clues to find
the problem.

See #11076, #10332, #4755

* ENH: Be more forgiving of malformed WAV chunks

Some of the metadata chunks at the end of WAV files are not the length
they say they are, but if we already successfully read the format and
data chunks, we can ignore them.

See #10332, #10870

* MAINT: Wave formats from int constants to IntEnum

These can be accessed by both value and name, and can output the name
as a string for error messages.

* ENH: Add all wave formats for error messages

and also for a future info() function.  Also note how to find the
latest list.

* MAINT: Set literals in membership tests

These are faster and preferred in Python 3, and SciPy now
requires >= 3.6

* STY: PEP8 and whitespace formatting

Break out filenames for PEP8 line length and code similarity

* MAINT: Use pytest.raises/warns instead of aliases

* ENH: Include format name in "Unknown format" error

* TST: Add .ani file (RIFF but not WAVE) for tests

These are RIFF format, but not WAVE form type.  Chosen because it's
small and public domain, from:
http://www.rw-designer.com/cursor-detail/113996

* TST: Add ulaw format WAV file for tests

Generated in Ocenaudio

ulaw format is not currently supported by SciPy

9 samples of 1 kHz sine at 8 kHz sample rate

* TST: Add unrecognized formats, error messages

Test that error messages include the more-helpful output strings, and
add tests for other error conditions.

Add comments describing each test file