BUG: stats: wilcoxon returned p > 1 for certain inputs.
When the W statistic associated with an input array occurred
at the center of the distribution and the length of the support
of the distribution was odd, the two-sided calculation of the
p-value counted the middle value of the distribution twice,
resulting in a p-value greater than 1.

The code now handles this case explicitly.