MAINT: io: Change the encoding comment of test_mio.py to utf-8.
The file test_mio.py had a comment indicating that its encoding
was latin-1.  The only non-ASCII character in the file is ö (in
the string 'Schrödinger').  Inspection of the raw bytes of this
file shows that this character is represented as the two byte
sequence 0xC3 0xB6, which is the UTF-8 encoding of ö. Apparently
the encoding of the file is actually UTF-8, not latin-1, so the
coding comment is changed accordingly.