ENH: add solver for minimum weight full bipartite matching
This adds a function
`scipy.sparse.csgraph.minimum_weight_full_bipartite_matching` which solves
the rectangular linear assignment problem for sparse inputs: Finding
a matching in a bipartite graph minimizing the sum of weights of the
edges in the matching. This is a generalization of the problem solved
by `maximum_bipartite_matching`, and a special case of the minimum-cost
flow problem.

The solver is a port of the Pascal implementation of the 1987 algorithm
LAPJVsp by Jonker and Volgenant. We follow the original implementation
closely, opting to having a solution that's close to the original
implementation rather than completely optimized Cython code.