BUG: Fix problem with Scipy.integrate.solve_bvp for big problems (#12556)
* BUG: Fix problem with Scipy.integrate.solve_bvp for big problems with parameters.

This commit solves fatal errors when solve_bvp is used for a big problem, i.e. size > 50,
and the unknowns include parameters that need solving.
This commit includes tests, with and without Jacobians for both the function and the boundary conditions.
The new tests adapt already existing tests to make them easier to compare and understand.