ENH: Read arbitrary bit depth (including 24-bit) WAVs (#12287)
* ENH: Read arbitrary bit depth WAVs

3-byte code is based on https://github.com/perimosocordiae/scipy/commit/9dba3f35258273fd81bf9bf7ea030c6e14780389
which is based on https://gist.github.com/josephernest/3f22c5ed5dabf1815f16efa8fa53d476
which is based on https://github.com/WarrenWeckesser/wavio/blob/master/wavio.py#L60

* MAINT: Add _raise_bad_format()

* Remove float16 support

This probably doesn't exist, and the tests say 8- or 16-bit float is not
expected.

The original bit depth check included bit depths for both int and float
formats in the same conditional.

* DOC: Add note about supported float types

* TST: Specify supported types directly

in test_write_roundtrip instead of generating many different
combinations automatically and then rejecting the majority of them

* Simplify 24-bit rearrangement

* Try to add support for all bit depths 1-64

It's doubtful that anything over 32 exists, but the previous code
supported 64-bit

* Remove platform-dependent int and float tests

This was causing test failures for file types that don't exist in the
real world, and files that are interchanged between platforms shouldn't
use platform-dependent types anyway.  See
https://github.com/scipy/scipy/issues/6090 and
https://github.com/scipy/scipy/pull/6184#issuecomment-636357138

* TST: Add 24-bit test with full-scale values

Odd number of channels, samples, and bytes per sample.

Data chunk is then also odd-length, so it requires a pad byte.

* TST: Add 5-bit example test

* TST: Add 12-bit example test

* TST: Add 64-bit example test

* TST: Add 36-, 45-, 53-bit example tests

To test 5-, 6-, and 7-byte ranges.

These probably don't exist, but testing for completeness

* TST: Add 20-bit WAV with extra data

Some software writes extra bits to fill the container size even if it
exceeds the specified bit depth.

* TST: WAV formats that can't be memory mapped

* DOC: Clarify integer, float, A-law are all PCM

* MAINT: Use np.void instead of uint8 for raw bytes

* TST: Remove "del data" from non-mmaped tests

* Specify int vs float in error messages

* Use number of bytes in fromfile instead of items

When loading the file with fromfile and a data type that needs
reshaping later, we need to use the number of raw bytes instead of
items.