MAINT: move common linear assignment tests to _lib
This moves the commons tests for scipy.optimize.linear_sum_assignment
and scipy.sparse.csgraph.min_weight_full_bipartite_matching to a common
test module in _lib/tests.

At the same time, we introduce a new test asserting that the two methods
produce the same result on a variety of test inputs.