REL: put upper bound on NumPy version range we support
Rationale: future releases of NumPy are likely to break an
already released version of SciPy, so really we should put an
upper bound. In this case one more than the already released
NumPy versions (we're at 1.19.2 now, and support all 1.20.x)
to avoid the next NumPy release making SciPy un-installable.